requirements.txt
confluent_kafka==2.2.0
Flask==3.0.0
nltk==3.8.1
numpy==1.26.1
pandas==2.1.2
praw==7.7.1
pymongo==4.5.0
python-dotenv==1.0.0
scikit_learn==1.3.1
wordcloud==1.9.2
